Move folder for related import declarations endpoint
